# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/AddControlClass": {
            "post": {
                "description": "添加一个控制模块",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "添加一个控制模块",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.AddControlClassReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.AddControlClassResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/AddControlDetail": {
            "post": {
                "description": "控制模块-键",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "控制模块-键"
                ],
                "summary": "添加一个控制模块-键",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.AddControlDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.AddControlDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/DeleteControlClass": {
            "post": {
                "description": "删除控制模块",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "删除控制模块",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.DeleteControlClassReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.DeleteControlClassResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/DeleteControlDetail": {
            "post": {
                "description": "删除控制模块",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "控制模块-键"
                ],
                "summary": "删除控制模块",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.DeleteControlDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.DeleteControlDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/ExecControlDetail": {
            "post": {
                "description": "执行key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "控制模块-键"
                ],
                "summary": "执行key",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.ExecControlDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.ExecControlDetailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/GetControlClassInfo": {
            "post": {
                "description": "获取控制模块详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "获取控制模块详情",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.GetControlClassInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.GetControlClassInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/GetControlClassList": {
            "post": {
                "description": "获取控制模块列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "获取控制模块列表",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.GetControlClassListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.GetControlClassListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/GetControlDetailList": {
            "post": {
                "description": "获取键列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "控制模块-键"
                ],
                "summary": "获取键列表",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.GetControlDetailListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.GetControlDetailListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/GetNowIconList": {
            "post": {
                "description": "获取当前图标列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取当前图标列表",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.GetNowIconListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.GetNowIconListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/GetNowMousePosition": {
            "post": {
                "description": "执行key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "控制模块-键"
                ],
                "summary": "获取当前鼠标位置",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.GetNowMousePositionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.GetNowMousePositionResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/GetSystemConfig": {
            "post": {
                "description": "获取当前系统设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置"
                ],
                "summary": "获取当前系统设置",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hd.GetSystemConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/hd.GinResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hd.GetSystemConfigResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "hd.AddControlClassReq": {
            "type": "object",
            "properties": {
                "control_id": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                },
                "control_name": {
                    "type": "string",
                    "example": ""
                },
                "control_order": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                },
                "key_height": {
                    "type": "string"
                },
                "key_width": {
                    "type": "string"
                }
            }
        },
        "hd.AddControlClassResp": {
            "type": "object"
        },
        "hd.AddControlDetailReq": {
            "type": "object",
            "properties": {
                "control_id": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                },
                "detail": {
                    "$ref": "#/definitions/config.ControlDetailExt"
                }
            }
        },
        "hd.AddControlDetailResp": {
            "type": "object"
        },
        "hd.DeleteControlClassReq": {
            "type": "object",
            "properties": {
                "control_id": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                }
            }
        },
        "hd.DeleteControlClassResp": {
            "type": "object"
        },
        "hd.DeleteControlDetailReq": {
            "type": "object",
            "properties": {
                "control_id": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                },
                "detail_id": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                }
            }
        },
        "hd.DeleteControlDetailResp": {
            "type": "object"
        },
        "hd.ExecControlDetailReq": {
            "type": "object",
            "properties": {
                "control_id": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                },
                "detail_id": {
                    "type": "integer",
                    "default": 0,
                    "example": 0
                }
            }
        },
        "hd.ExecControlDetailResp": {
            "type": "object"
        },
...
        "hd.GinResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/hd.STATE_CODE"
                }
            }
        },
...
    }
}